fix: cast adaptive timesteps to weight dtype in Anima model_fn
The training path returns timesteps already cast to weight_dtype from
flux_train_utils.get_noisy_model_input_and_timesteps, so timesteps/1000.0
stays bf16. Anima's sinusoidal embedder inherits the timestep dtype
(emb.to(in_dtype)), and with AdaLN-LoRA the raw sinusoidal embedding is fed
directly to the AdaLN linear layers, which run with autocast explicitly
disabled (enabled=use_fp32). The adaptive model_fn passed float32 timesteps,
producing a float32 embedding that crashed against bf16 AdaLN weights with
'expected mat1 and mat2 to have the same dtype'. Cast the scaled timesteps to
wdtype to exactly mirror the training path; outer autocast alone cannot help
since the AdaLN block disables it.

+++ anima_train_network.py
@@ -112,8 +112,13 @@ class AnimaNetworkTrainer(train_network.NetworkTrainer):
             else:
                 et5m = None
 
-            # Scale timesteps to [0, 1] range for Anima (Anima expects timesteps / 1000)
-            anima_ts = timesteps.float() / 1000.0
+            # Scale timesteps to [0, 1] range for Anima (Anima expects timesteps / 1000).
+            # IMPORTANT: cast to wdtype (bf16/fp16) to match the training path, where
+            # flux_train_utils.get_noisy_model_input_and_timesteps returns timesteps in
+            # weight_dtype. The sinusoidal embedder inherits the timestep dtype, and the
+            # AdaLN block runs with autocast explicitly disabled (enabled=use_fp32), so a
+            # float32 embedding would crash against bf16 AdaLN weights.
+            anima_ts = (timesteps.float() / 1000.0).to(wdtype)
 
             # 4D to 5D: [N, C, H, W] -> [N, C, 1, H, W]
             x_5d = noisy_latents.to(wdtype).unsqueeze(2)
